Have snackbar only in DeleteDeviceButton
